Some random objects  around
