+++ Code/Assets/Scripts/ObjBehaviour/StarBehaviour.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class StarBehaviour : MonoBehaviour {
+    private void OnCollisionEnter2D(Collision2D collision)
+    {
+        if (collision.gameObject.tag != "Flushi")
+            Destroy(collision.gameObject);
+    }
+       
+}
