+++ Code/Assets/Scripts/ObjBehaviour/StarBehaviour.cs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 859ac7201dc54d74da9d40a4d89e23e3
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
